Correct scripts

+++ Assets/Sources/UI/Scripts/CardWand.cs
@@ -6,10 +6,10 @@ public class CardWand : Card
 {
     [Space]
     [Header("Характеристики палочки")]
-    [SerializeField] private float _damage;
+    [SerializeField] private int _damage;
     [SerializeField] private float _attackSpeed;
 
-    public float Damage => _damage;
+    public int Damage => _damage;
     public float AttackSpeed => _attackSpeed;
 
     public CardWand ReturnCard() => this;
